Find approx. Bayes factors (ABFs)

Wakefield's log asymptotic Bayes factor (lABF) with prior standard deviation of effect size as a parameter

approx.bf.p(pvals, f, type, N, s, W = 0.2)

Arguments

pvals
P-values

f
Minor allele frequencies

type
Type of experiment ('quant' or 'cc')

N
Total sample size

s
Proportion of cases (N1/N0+N1), ignored if type=='quant'

W
Prior for the standard deviation of the effect size parameter beta (W=0.2 default)

Value

data.frame containing lABF and intermediate calculations

Details

([Wakefield et al. 2009](https://onlinelibrary.wiley.com/doi/abs/10.1002/gepi.20359) This function converts p-values to log ABFs, also reporting intermediate calculations

Examples

set.seed(1)
nsnps = 100
N0 = 5000
N1 = 5000
z_scores <- rnorm(nsnps, 0, 3)
p_values <- 2 * pnorm( - abs ( z_scores ) )

nhaps <- 1000
lag <- 5
maf.tmp <- runif(nsnps+lag, 0.05, 0.5) # common SNPs
laghaps <- do.call("cbind", lapply(maf.tmp, function(f) rbinom(nhaps,1,f)))
haps <- laghaps[,1:nsnps]
for(j in 1:lag)
   haps <- haps + laghaps[,(1:nsnps)+j]
haps <- round(haps/matrix(apply(haps,2,max),nhaps,nsnps,byrow=TRUE))
maf <- colMeans(haps)

approx.bf.p(pvals = p_values, f = maf, type = "cc", N = N0+N1, s = N1/(N0+N1))

#>                V           z         r        lABF
#> 1   0.0021835015 1.879361432 0.9482380  0.19403835
#> 2   0.0018285714 0.550929973 0.9562842 -1.41989479
#> 3   0.0018540664 2.506885837 0.9557016  1.44463847
#> 4   0.0022222222 4.785842406 0.9473684  9.37717987
#> 5   0.0017234842 0.988523315 0.9586927 -1.12495123
#> 6   0.0046538685 2.461405152 0.8957791  1.58292449
#> 7   0.0014446274 1.462287157 0.9651432 -0.64637841
#> 8   0.0013101352 2.214974115 0.9682854  0.64976869
#> 9   0.0009383680 1.727344055 0.9770785 -0.43017727
#> 10  0.0008354533 0.916165161 0.9795410 -1.53357267
#> 11  0.0009004507 4.535343505 0.9779843  8.15024600
#> 12  0.0025715204 1.169529709 0.9395953 -0.76075499
#> 13  0.0014880952 1.863721742 0.9641320  0.01048155
#> 14  0.0013491726 6.644099662 0.9673712 19.64056653
#> 15  0.0019651191 3.374792754 0.9531726  3.89730596
#> 16  0.0010123558 0.134800827 0.9753158 -1.84193528
#> 17  0.0010753208 0.048570789 0.9738208 -1.82024552
#> 18  0.0047546596 2.831508632 0.8937617  2.46180562
#> 19  0.0017344399 2.463663585 0.9584410  1.31837413
#> 20  0.0016812939 1.781703964 0.9596631 -0.08203398
#> 21  0.0014806150 2.756932115 0.9643059  1.99830324
#> 22  0.0015600746 2.346408902 0.9624622  1.00827944
#> 23  0.0017234842 0.223694950 0.9586927 -1.56937188
#> 24  0.0015516385 5.968055088 0.9626576 15.50000291
#> 25  0.0013209691 1.859477244 0.9680315 -0.04794245
#> 26  0.0016711370 0.168386219 0.9598970 -1.59454384
#> 27  0.0119681647 0.467386520 0.7697020 -0.65011980
#> 28  0.0018939394 4.412257152 0.9547920  7.74571095
#> 29  0.0013491726 1.434450165 0.9673712 -0.71602605
#> 30  0.0014659425 1.253824681 0.9646471 -0.91293761
#> 31  0.0014806150 4.076038655 0.9643059  6.34414836
#> 32  0.0012453378 0.308363182 0.9698066 -1.70395723
#> 33  0.0036605901 1.163014835 0.9161580 -0.61981147
#> 34  0.0037832930 0.161415122 0.9135905 -1.21242703
#> 35  0.0038480038 4.131178670 0.9122422  6.56786610
#> 36  0.0011505891 1.244983690 0.9720395 -1.03515902
#> 37  0.0011731857 1.182869861 0.9715061 -1.09937571
#> 38  0.0010666667 0.177940190 0.9740260 -1.80990897
#> 39  0.0013209691 3.300076116 0.9680315  3.54967247
#> 40  0.0013669045 2.289527245 0.9669566  0.82939551
#> 41  0.0014659425 0.493570789 0.9646471 -1.55368717
#> 42  0.0013790821 0.760085040 0.9666720 -1.42144133
#> 43  0.0019801196 2.090890126 0.9528320  0.55578619
#> 44  0.0031994369 1.669989596 0.9259380 -0.01026817
#> 45  0.0015270090 2.066267084 0.9632285  0.40471612
#> 46  0.0012227630 2.122485471 0.9703377  0.42671993
#> 47  0.0009542030 1.093745886 0.9767007 -1.29546282
#> 48  0.0028828829 2.305598774 0.9327731  1.12937000
#> 49  0.0018540664 0.337038636 0.9557016 -1.50412235
#> 50  0.0019216726 2.643323179 0.9541604  1.79213098
#> 51  0.0011028033 1.194317641 0.9731696 -1.11504849
#> 52  0.0013729569 1.836079180 0.9668151 -0.07317322
#> 53  0.0016916037 1.023359074 0.9594258 -1.09992542
#> 54  0.0017234842 3.388089288 0.9586927  3.90913012
#> 55  0.0014659425 4.299071105 0.9646471  7.24312253
#> 56  0.0014174344 5.941199696 0.9657769 15.35749860
#> 57  0.0012361627 1.101664429 0.9700224 -1.16501095
#> 58  0.0012642944 3.132403879 0.9693611  3.01292138
#> 59  0.0037832930 1.709158882 0.9135905  0.11007251
#> 60  0.0028480292 0.405163812 0.9335319 -1.27889298
#> 61  0.0014174344 7.204853282 0.9657769 23.37926876
#> 62  0.0018939394 0.117720008 0.9547920 -1.54162533
#> 63  0.0014880952 2.069218087 0.9641320  0.40008959
#> 64  0.0012183904 0.084006476 0.9704406 -1.75725270
#> 65  0.0017568672 2.229819627 0.9579263  0.79728462
#> 66  0.0030321867 0.566376899 0.9295368 -1.17724247
#> 67  0.0034390261 5.414875887 0.9208310 12.23170244
#> 68  0.0024178242 4.396664585 0.9429998  7.68205376
#> 69  0.0045574697 0.459760015 0.8977170 -1.04512673
#> 70  0.0057630244 6.517835011 0.8740681 17.53014929
#> 71  0.0054710581 1.426528587 0.8796804 -0.16373428
#> 72  0.0039847781 2.129839293 0.9094055  0.86194867
#> 73  0.0034916811 1.832179060 0.9197161  0.28259518
#> 74  0.0033386752 2.802292895 0.9229631  2.34220781
#> 75  0.0119681647 3.760900201 0.7697020  4.70928446
#> 76  0.0030321867 0.874338707 0.9295368 -0.97103153
#> 77  0.0022833396 1.329875620 0.9459991 -0.62284462
#> 78  0.0014108452 0.003316055 0.9659305 -1.68967163
#> 79  0.0013048102 0.223023972 0.9684102 -1.70337648
#> 80  0.0016916037 1.768562839 0.9594258 -0.10185840
#> 81  0.0025715204 1.706006198 0.9395953 -0.03601817
#> 82  0.0018540664 0.405535845 0.9557016 -1.47981681
#> 83  0.0023484101 3.534260990 0.9445455  4.45306308
#> 84  0.0043767507 4.570700401 0.9013729  8.25722214
#> 85  0.0059215396 1.781838563 0.8710509  0.35860210
#> 86  0.0052083333 0.998851114 0.8847926 -0.63913039
#> 87  0.0059215396 3.189299512 0.8710509  3.40583561
#> 88  0.0033880503 0.912551771 0.9219128 -0.89110292
#> 89  0.0027488008 1.110056430 0.9356988 -0.79559236
#> 90  0.0018540664 0.801296372 0.9557016 -1.25158735
#> 91  0.0013101352 1.627560093 0.9682854 -0.44301809
#> 92  0.0060893923 3.623603418 0.8678787  4.68582385
#> 93  0.0013433999 3.481207847 0.9675063  4.14915702
#> 94  0.0025715204 2.100640949 0.9395953  0.66972841
#> 95  0.0018804238 4.760500364 0.9551002  9.27075308
#> 96  0.0016812939 1.675459277 0.9596631 -0.25827862
#> 97  0.0032908268 3.829776625 0.9239833  5.48771772
#> 98  0.0022625458 1.719796243 0.9464645 -0.06402667
#> 99  0.0016916037 3.673837845 0.9594258  4.87241414
#> 100 0.0013155212 1.420201909 0.9681592 -0.74712700