Calculate ABFs from Z scores

Calculate ABFs from Z scores

bf_func(z, V, W = 0.2)

Arguments

z
Vector of Z-scores

V
Variance of the estimated effect size

W
Prior for the standard deviation of the effect size parameter, beta (default 0.2)

Value

ABFs

Details

Note, z and V should both be vectors or both be matrices

Examples

set.seed(1)
nsnps = 100
N0 = 5000
N1 = 5000
z_scores <- rnorm(nsnps, 0, 3)

## generate example LD matrix and MAFs
library(mvtnorm)
nsamples = 1000

simx <- function(nsnps, nsamples, S, maf=0.1) {
    mu <- rep(0,nsnps)
    rawvars <- rmvnorm(n=nsamples, mean=mu, sigma=S)
    pvars <- pnorm(rawvars)
    x <- qbinom(1-pvars, 1, maf)
}

S <- (1 - (abs(outer(1:nsnps,1:nsnps,`-`))/nsnps))^4
X <- simx(nsnps,nsamples,S)
maf <- colMeans(X)

varbeta = Var.data.cc(f = maf, N = N0 + N1, s = 0.5)

bf_func(z_scores, V = varbeta)

#>   [1]  0.259944111 -1.270473777  1.552208379  9.341047932 -0.944268001
#>   [6]  1.427301994 -0.447969805  0.864423167 -0.029378459 -1.027612048
#>  [11]  8.268242890 -0.832049849  0.197976322 19.374402884  3.934380483
#>  [16] -1.419097192 -1.466868876  2.337827252  1.377876401  0.027985426
#>  [21]  2.130632955  1.123928032 -1.488355879 15.399313866  0.194207970
#>  [26] -1.467106578 -1.389007422  7.748404451 -0.540087117 -0.800915466
#>  [31]  6.386598347 -1.502846526 -0.892703013 -1.521809335  6.607155676
#>  [36] -0.767748454 -0.860427213 -1.522683876  3.670287713  0.962190427
#>  [41] -1.373099565 -1.248220640  0.595415202 -0.201508038  0.509939166
#>  [46]  0.645967635 -0.921022804  1.022248367 -1.442733815  1.821801820
#>  [51] -0.811824791  0.101161970 -0.980104986  3.966937123  7.280874246
#>  [56] 15.261456261 -0.916538268  3.171484247 -0.109276062 -1.406782972
#>  [61] 23.052184326 -1.457147250  0.546274685 -1.460365661  0.867743410
#>  [66] -1.370589535 12.432580567  7.683014958 -1.380330779 18.688532357
#>  [71] -0.492936744  0.667016367  0.117881615  2.252514493  5.233904471
#>  [76] -1.118100815 -0.638275707 -1.492709470 -1.448658583 -0.001030821
#>  [81] -0.086383093 -1.354807751  4.451673295  8.419554766  0.024753811
#>  [86] -0.999623072  3.342014100 -1.065486305 -0.864149744 -1.138494311
#>  [91] -0.217452821  4.755081967  4.266740749  0.637898200  9.251555061
#>  [96] -0.124132598  5.477281671 -0.060386708  4.930438044 -0.464484167