Corrected credible set with desired coverage of the CV

Corrected credible set with desired coverage of the CV

corrected_cs(z, f, N0, N1, Sigma, lower = 0, upper = 1, desired.cov,
  acc = 0.005, max.iter = 20)

Arguments

z
Z-scores

f
Minor allele frequencies

N0
Number of controls

N1
Number of cases

Sigma
Correlation matrix of SNPs

lower
Lower threshold (default = 0)

upper
Upper threshold (default = 1)

desired.cov
The desired coverage of the causal variant in the credible set

acc
Accuracy of corrected coverage to desired coverage (default = 0.005)

max.iter
Maximum iterations (default = 20)

Value

List of variants in credible set, required threshold, the corrected coverage and the size of the credible set

Examples

# NOT RUN {
 # this is a long running example

# In this example, the function is used to find a corrected 95% credible set
# using Z-scores and MAFs, that is the smallest set of variants
# required such that the resultant credible set has coverage close to (/within
# some accuracy of) the "desired coverage" (here set to 0.95). Max.iter parameter
# defines the maximum number of iterations to try in the root bisection algorithm,
# this should be increased to ensure convergence to the desired coverage, but is set
# to 1 here for speed (and thus the resultant credible set will not be accurate).

set.seed(2)
nsnps = 200
N0 = 1000
N1 = 1000
z_scores <- rnorm(nsnps, 0, 1) # simulate a vector of Z-scores

## generate example LD matrix
library(mvtnorm)
nsamples = 1000

simx <- function(nsnps, nsamples, S, maf=0.1) {
    mu <- rep(0,nsnps)
    rawvars <- rmvnorm(n=nsamples, mean=mu, sigma=S)
    pvars <- pnorm(rawvars)
    x <- qbinom(1-pvars, 1, maf)
}

S <- (1 - (abs(outer(1:nsnps,1:nsnps,`-`))/nsnps))^4
X <- simx(nsnps,nsamples,S)
LD <- cor2(X)

names(z_scores) <- seq(1,length(z_scores))

corrected_cs(z = z_scores, f = maf, N0, N1, Sigma = LD, desired.cov = 0.9, max.iter = 1)
# max.iter set low for speed, should be set to at least
# the default to ensure convergence to desired coverage
# }