Credible set of variants from matrix of PPs

Quick credset function for matrix of posterior probabilities (using RCpp)

credsetC(pp, CV, thr)

Arguments

pp
Matrix of posterior probabilities of causality (one row per system)

CV
Vector of CV indices (one per system/row)

thr
Minimum threshold for credible set size

Value

Data.frame of claimed coverage (sum of posterior probabilities of variants in the set), binary covered indicator and number of variants (nvar).

Examples

set.seed(1)
nsnps <- 100

# simulate matrix of posterior probabilities
# 1 simulation per row

pp <- matrix(rnorm(nsnps*100, 0.3, 0.05), ncol = nsnps)
pp <- pp/rowSums(pp)

iCV <- rep(71, times = dim(pp)[1])

credsetC(pp, CV = iCV, thr = 0.9)

#>     claimed.cov covered nvar
#> 1     0.9028106       0   87
#> 2     0.9043994       1   87
#> 3     0.9066858       1   87
#> 4     0.9073203       1   87
#> 5     0.9016568       1   87
#> 6     0.9066431       1   87
#> 7     0.9023015       1   87
#> 8     0.9073503       1   88
#> 9     0.9051947       1   87
#> 10    0.9065068       1   87
#> 11    0.9003956       1   86
#> 12    0.9022983       1   87
#> 13    0.9037407       1   87
#> 14    0.9075461       1   87
#> 15    0.9006862       0   86
#> 16    0.9072241       1   86
#> 17    0.9044851       1   87
#> 18    0.9005781       0   86
#> 19    0.9013418       1   87
#> 20    0.9032404       1   87
#> 21    0.9058975       1   87
#> 22    0.9009720       1   87
#> 23    0.9045454       0   87
#> 24    0.9018952       1   86
#> 25    0.9014670       1   87
#> 26    0.9058481       1   87
#> 27    0.9077576       0   87
#> 28    0.9070294       1   87
#> 29    0.9043705       0   87
#> 30    0.9041687       1   87
#> 31    0.9049257       1   87
#> 32    0.9043269       1   86
#> 33    0.9004033       1   86
#> 34    0.9035822       1   87
#> 35    0.9029281       1   87
#> 36    0.9059044       1   87
#> 37    0.9007131       1   86
#> 38    0.9079439       1   87
#> 39    0.9053566       1   87
#> 40    0.9032632       1   87
#> 41    0.9024774       1   86
#> 42    0.9076526       1   87
#> 43    0.9057440       1   87
#> 44    0.9042658       1   87
#> 45    0.9074380       1   87
#> 46    0.9050979       1   87
#> 47    0.9026150       0   87
#> 48    0.9043232       1   87
#> 49    0.9002837       1   86
#> 50    0.9023159       0   87
#> 51    0.9007277       1   86
#> 52    0.9010689       1   87
#> 53    0.9038356       1   87
#> 54    0.9052344       1   87
#> 55    0.9068687       1   87
#> 56    0.9016287       1   86
#> 57    0.9046650       1   86
#> 58    0.9023235       1   87
#> 59    0.9034334       1   87
#> 60    0.9048217       1   87
#> 61    0.9062914       1   87
#> 62    0.9023418       0   87
#> 63    0.9056378       1   87
#> 64    0.9034765       1   87
#> 65    0.9016205       1   86
#> 66    0.9079594       1   87
#> 67    0.9013745       0   86
#> 68    0.9000213       1   86
#> 69    0.9083436       1   88
#> 70    0.9042942       1   87
#> 71    0.9072779       0   87
#> 72    0.9040201       1   87
#> 73    0.9049649       1   87
#> 74    0.9059295       1   87
#> 75    0.9022478       1   87
#> 76    0.9058106       1   87
#> 77    0.9030499       1   86
#> 78    0.9073147       1   87
#> 79    0.9067678       1   87
#> 80    0.9033786       0   87
#> 81    0.9000792       1   87
#> 82    0.9054505       1   87
#> 83    0.9012045       1   87
#> 84    0.9023899       0   87
#> 85    0.9014795       0   86
#> 86    0.9036947       1   87
#> 87    0.9078739       1   87
#> 88    0.9003364       1   87
#> 89    0.9002072       1   86
#> 90    0.9009011       1   87
#> 91    0.9045547       1   87
#> 92    0.9019096       1   87
#> 93    0.9007869       0   86
#> 94    0.9028258       1   87
#> 95    0.9053618       1   87
#> 96    0.9032200       1   87
#> 97    0.9004761       1   86
#> 98    0.9062791       1   87
#> 99    0.9073267       1   87
#> 100   0.9009760       1   86