Find PPs from Z-scores and MAFs

Posterior probabilities of causality from marginal Z-scores with prior SD as a parameter

z0_pp(z, f, type, N, s, W = 0.2)

Arguments

z
Marginal Z-scores of SNPs

f
Minor allele frequencies

type
Type of experiment ('quant' or 'cc')

N
Total sample size

s
Proportion of cases (N1/N0+N1), ignored if type=='quant'

W
Prior for the standard deviation of the effect size parameter beta

Value

Posterior probabilities of null model (no genetic effect) and causality for each SNP

Details

This function converts Z-scores to posterior probabilities of causality, including the null model of no genetic effects

Examples

set.seed(1)
nsnps = 100
N0 = 5000
N1 = 5000
z_scores <- rnorm(nsnps, 0, 3)

nhaps <- 1000
lag <- 5
maf.tmp <- runif(nsnps+lag, 0.05, 0.5) # common SNPs
laghaps <- do.call("cbind", lapply(maf.tmp, function(f) rbinom(nhaps,1,f)))
haps <- laghaps[,1:nsnps]
for(j in 1:lag)
   haps <- haps + laghaps[,(1:nsnps)+j]
haps <- round(haps/matrix(apply(haps,2,max),nhaps,nsnps,byrow=TRUE))
maf <- colMeans(haps)

res <- z0_pp(z = z_scores, f = maf, type = "cc", N = N0+N1, s = N1/(N0+N1))
sum(res)

#> [1] 1

res

#>   [1] 6.767241e-07 8.299391e-11 1.652433e-11 2.898511e-10 8.076699e-07
#>   [6] 2.219302e-11 3.328372e-10 3.581442e-11 1.309084e-10 4.445830e-11
#>  [11] 1.474872e-11 2.368009e-07 3.194367e-11 6.907621e-11 2.315081e-02
#>  [16] 3.367867e-09 1.083513e-11 1.107271e-11 8.015397e-10 2.554695e-10
#>  [21] 6.297230e-11 5.042298e-10 1.873556e-10 1.423007e-11 3.684191e-04
#>  [26] 6.515613e-11 1.387634e-11 3.568067e-11 1.580141e-07 3.340491e-11
#>  [31] 2.743422e-11 3.890497e-08 1.243819e-11 3.677865e-11 2.033415e-11
#>  [36] 4.865914e-08 2.427798e-11 2.276793e-11 1.118776e-11 2.378919e-09
#>  [41] 1.566673e-10 1.445502e-11 1.649880e-11 1.191657e-10 6.765767e-11
#>  [46] 1.024572e-10 1.047366e-10 1.871389e-11 2.114733e-10 1.518953e-11
#>  [51] 4.102882e-10 2.241388e-11 6.353276e-11 2.275542e-11 3.407925e-09
#>  [56] 9.559267e-08 3.194871e-04 2.132154e-11 1.390822e-09 7.630980e-11
#>  [61] 1.902656e-11 9.733384e-01 1.463043e-11 1.019843e-10 1.179265e-11
#>  [66] 1.517165e-10 2.106234e-11 1.402611e-05 1.482689e-07 2.403719e-11
#>  [71] 2.805586e-03 5.803200e-11 1.618512e-10 9.067884e-11 7.111879e-10
#>  [76] 7.585648e-09 2.588587e-11 3.666726e-11 1.261715e-11 1.244542e-11
#>  [81] 6.173620e-11 6.593772e-11 1.556325e-11 5.871069e-09 2.635375e-07
#>  [86] 9.783976e-11 3.607494e-11 2.060213e-09 2.803983e-11 3.084999e-11
#>  [91] 1.955325e-11 4.389107e-11 7.409756e-09 4.332439e-09 1.335475e-10
#>  [96] 7.261283e-07 5.279678e-11 1.652198e-08 6.411653e-11 8.929744e-09
#> [101] 3.238197e-11