Simulate posterior probabilities of causality from joint Z-score vector

Simulate nrep marginal Z-scores from joint Z-scores and convert these to posterior probabilities of causality

zj_pp(Zj, V, nrep = 5000, W = 0.2, Sigma)

Arguments

Zj
Vector of joint Z-scores (0s except at CV)

V
Variance of the estimated effect size (can be obtained using Var.beta.cc function)

nrep
Number of posterior probability systems to simulate

W
Prior for the standard deviation of the effect size parameter, beta

Sigma
SNP correlation matrix

Value

Matrix of simulated posterior probabilties, one simulation per row

Details

Does not include posterior probabilities for null model

Examples

set.seed(1)
nsnps <- 100
Zj <- rep(0, nsnps)
iCV <- 4 # index of CV
mu <- 5 # true effect at CV
Zj[iCV] <- mu

## generate example LD matrix (https://chr1swallace.github.io/simGWAS/articles/intro.html)
nhaps <- 1000
lag <- 5
maf.tmp <- runif(nsnps+lag, 0.05, 0.5) # common SNPs
laghaps <- do.call("cbind", lapply(maf.tmp, function(f) rbinom(nhaps,1,f)))
haps <- laghaps[,1:nsnps]
for(j in 1:lag)
   haps <- haps + laghaps[,(1:nsnps)+j]
haps <- round(haps/matrix(apply(haps,2,max),nhaps,nsnps,byrow=TRUE))
LD <- cor2(haps)
maf <- colMeans(haps)

## generate V (variance of estimated effect sizes)
varbeta <- Var.data.cc(f = maf, N = 5000, s = 0.5)

res <- zj_pp(Zj, V = varbeta, nrep = 5, W = 0.2, Sigma = LD)

res[c(1:5), c(1:5)]

#>              [,1]         [,2]         [,3]      [,4]         [,5]
#> [1,] 2.500106e-06 2.723245e-06 0.0002295946 0.9996125 7.012648e-05
#> [2,] 4.021839e-05 2.056739e-04 0.0058773031 0.9937625 2.442096e-05
#> [3,] 1.234702e-03 8.079721e-04 0.0179892575 0.9790939 9.782115e-05
#> [4,] 2.933681e-06 3.387701e-06 0.0001042174 0.9998139 1.256225e-05
#> [5,] 2.112681e-04 1.027380e-03 0.0542606973 0.9271044 1.470160e-04